% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl-genes.R
\name{ensembl_genes}
\alias{ensembl_genes}
\title{Generate a table of gene symbols and Ensembl IDs}
\usage{
ensembl_genes(x)
}
\arguments{
\item{x}{A list of data frames returned by \code{msigdb_sqlite()}.}
}
\value{
A data frame with gene details.
}
\description{
Simplify the table of Ensembl ID mappings provided by MSigDB.
Although these are the Ensembl canonical IDs for each gene, there are some genes with more than ten IDs.
This function additionally reduces the number of multi-mapping IDs based on those actually appearing in MSigDB.
}
