% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-set-details.R
\name{gene_set_details}
\alias{gene_set_details}
\title{Generate a table of gene sets with details}
\usage{
gene_set_details(x)
}
\arguments{
\item{x}{A list of data frames returned by \code{msigdb_sqlite()}.}
}
\value{
A data frame with gene set details.
}
\description{
Convert the tables derived from the MSigDB SQLite database to a single table of gene set information.
The output includes the full name, description, source publication, and other details for each gene set.
}
