% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-set-genes.R
\name{gene_set_genes}
\alias{gene_set_genes}
\title{Generate a table of gene details}
\usage{
gene_set_genes(x)
}
\arguments{
\item{x}{A list of data frames returned by \code{msigdb_sqlite()}.}
}
\value{
A data frame with gene details.
}
\description{
Convert the MSigDB SQLite database tables to a single table of gene information.
}
