% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-set-members.R
\name{gene_set_members}
\alias{gene_set_members}
\title{Generate a table of gene set members}
\usage{
gene_set_members(x)
}
\arguments{
\item{x}{A list of data frames returned by \code{msigdb_sqlite()}.}
}
\value{
A data frame with genes belonging to each gene set.
}
\description{
Convert the tables derived from the MSigDB SQLite database to a single table of member genes belonging to each gene set.
The output includes gene symbols, NCBI (formerly Entrez) IDs, and Ensembl IDs for each gene set.
}
