% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdb-ensembl.R
\name{msigdb_ensembl}
\alias{msigdb_ensembl}
\title{Read MSigDB Ensembl gene ID mappings}
\usage{
msigdb_ensembl(x)
}
\arguments{
\item{x}{A list of data frames returned by \code{msigdb_sqlite()}.}
}
\value{
A data frame with Ensembl gene IDs.
}
\description{
Download the Ensembl gene ID mappings compiled for MSigDB.
}
\details{
MSigDB versions and the corresponding Ensembl releases for gene annotation:
\itemize{
\item 2023.1 - 109 (February 2023)
\item 2023.2 - 110 (July 2023)
\item 2024.1 - 112 (May 2024)
}
}
