% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdb-sqlite.R
\name{msigdb_sqlite}
\alias{msigdb_sqlite}
\title{Read MSigDB SQLite database}
\usage{
msigdb_sqlite(x)
}
\arguments{
\item{x}{MSigDB version, such as \verb{2023.1.Hs}.}
}
\value{
A list of data frames.
}
\description{
Download the MSigDB SQLite database and extract the relevant tables as data frames.
Each database file holds one MSigDB release for one resource (human or mouse).
}
\references{
MSigDB SQLite database documentation: \url{https://docs.gsea-msigdb.org/#MSigDB/MSigDB_SQLite_Database/}
}
