% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbdf.R
\name{msigdbdf}
\alias{msigdbdf}
\title{Retrieve the MSigDB gene sets}
\usage{
msigdbdf(target_species = c("HS", "MM"))
}
\arguments{
\item{target_species}{Species abbreviation for human or mouse databases (\code{"HS"} or \code{"MM"}).}
}
\value{
A tibble (a data frame with class \code{\link{tbl_df}}) of gene sets with one gene per row.
}
\description{
Retrieve a data frame of MSigDB gene sets and their member genes.
Starting with release 2022.1, MSigDB was split into human and mouse resources, each one provided in the approved gene symbols of its respective species.
The MSigDB versioning convention is in the format \code{Year.Release.Species}.
The species referenced in this function is the one specified in the release version.
}
