% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbrdata-package.R
\docType{package}
\name{msigdbrdata-package}
\alias{msigdbrdata-package}
\title{msigdbrdata: Data for msigdbr}
\description{
Data for the msigdbr R package. The data was split from the msigdbr package to comply with CRAN package size limits.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/igordot/msigdbrdata}
  \item Report bugs at \url{https://github.com/igordot/msigdbrdata/issues}
}

}
\author{
\strong{Maintainer}: Igor Dolgalev \email{igor.dolgalev@nyumc.org} (\href{https://orcid.org/0000-0003-4451-126X}{ORCID})

}
\keyword{internal}
