% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbrdata.R
\name{msigdbrdata}
\alias{msigdbrdata}
\title{Retrieve the gene sets data frame}
\usage{
msigdbrdata(species = c("Hs", "Mm"))
}
\arguments{
\item{species}{Species abbreviation for human or mouse databases (\code{"Hs"} or \code{"Mm"}).}
}
\value{
A tibble (a data frame with class \code{\link{tbl_df}}) of gene sets with one gene per row.
}
\description{
Retrieve a data frame of gene sets and their member genes.
Starting with release 2022.1, MSigDB is split into human and mouse resources.
Each one is provided in the approved gene symbols of its respective species.
The versioning convention of MSigDB is in the format \code{Year.Release.Species}.
}
