% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-genesets.R
\name{prepare_genesets}
\alias{prepare_genesets}
\title{Generate a table of gene set details}
\usage{
prepare_genesets(x)
}
\arguments{
\item{x}{A list of data frames returned by \code{read_msigdb_sqlite()}.}
}
\value{
A data frame with gene set details.
}
\description{
Convert the MSigDB SQLite database tables to a single table of gene set information.
}
