% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-msigdb-sqlite.R
\name{read_msigdb_sqlite}
\alias{read_msigdb_sqlite}
\title{Read MSigDB SQLite database}
\usage{
read_msigdb_sqlite(x)
}
\arguments{
\item{x}{MSigDB version, such as \verb{2023.1.Hs}.}
}
\value{
A list of data frames.
}
\description{
Download the MSigDB SQLite database and extract the relevant tables as data frames.
}
